+++ lightstep/resource_metric_condition.go
@@ -1000,36 +1000,6 @@ func setResourceDataFromUnifiedCondition(project string, c client.UnifiedConditi
 	return nil
 }
 
-func getQueriesFromUnifiedConditionResourceData(
-	queriesIn []client.MetricQueryWithAttributes,
-	conditionID string,
-) ([]interface{}, error) {
-	var queries []interface{}
-	for _, q := range queriesIn {
-		if q.Type != "tql" {
-			return nil, fmt.Errorf(
-				"cannot convert query from condition %v\n\n"+
-					"Query is of type '%v' but must be of type 'tql' for use with the resource\n"+
-					"type lightstep_condition.\n"+
-					"\n"+
-					"Try using the lightstep_metrics_condition resource type for this condition\n"+
-					"or convert all queries in the condition to query string format. ",
-				conditionID,
-				q.Type,
-			)
-		}
-
-		qs := map[string]interface{}{
-			"hidden":       q.Hidden,
-			"display":      q.Display,
-			"query_name":   q.Name,
-			"query_string": q.TQLQuery,
-		}
-		queries = append(queries, qs)
-	}
-	return queries, nil
-}
-
 func getIncludeExcludeFilters(filters []client.LabelFilter) ([]interface{}, []interface{}, []interface{}) {
 	var includeFilters []interface{}
 	var excludeFilters []interface{}
